Remove ResultMerger, KifiSearchResult
