+++ kifi-backend/modules/search/app/com/keepit/search/result/ResultUtil.scala
@@ -2,35 +2,13 @@ package com.keepit.search.result
 
 import com.keepit.search.engine.Visibility
 import com.keepit.search.engine.uri.{ UriShardHit, UriSearchResult }
-import play.api.libs.json._
 import com.keepit.common.db.{ Id, ExternalId }
 import org.joda.time.DateTime
 import com.keepit.search.ArticleHit
 import com.keepit.search.ArticleSearchResult
-import com.keepit.model.{ NormalizedURI, URISummary }
+import com.keepit.model.{ NormalizedURI }
 
 object ResultUtil {
-
-  private def jsonToKifiSearchHit(json: JsObject): KifiSearchHit = {
-    val uriSummaryJson = (json \ "uriSummary")
-    // All the fields of URISummary are nullable so we want to distinguish between "null" and an empty URI summary
-    val uriSummary = if (!uriSummaryJson.isInstanceOf[JsUndefined]) {
-      uriSummaryJson.asOpt[URISummary].map("uriSummary" -> Json.toJson(_)).toList
-    } else List()
-    KifiSearchHit(JsObject(List(
-      "count" -> (json \ "bookmarkCount"),
-      "bookmark" -> (json \ "bookmark"),
-      "users" -> (json \ "basicUsers"),
-      "score" -> (json \ "score"),
-      "isMyBookmark" -> (json \ "isMyBookmark"),
-      "isPrivate" -> (json \ "isPrivate")
-    ) ++ uriSummary))
-  }
-
-  def toKifiSearchHits(hits: Seq[DetailedSearchHit], sanitize: Boolean): Seq[KifiSearchHit] = {
-    if (sanitize) hits.map { h => jsonToKifiSearchHit(h.sanitized.json) }
-    else hits.map { h => jsonToKifiSearchHit(h.json) }
-  }
 
   def toArticleSearchResult(
     res: UriSearchResult,
