+++ kifi-backend/modules/search/test/com/keepit/search/tracking/SearchAnalyticsTest.scala
@@ -1,9 +1,7 @@
 package com.keepit.search.tracking
 
-import com.keepit.common.db.{ ExternalId, Id }
+import com.keepit.common.db.{ ExternalId }
 import com.keepit.model._
-import com.keepit.search.result.{ BasicSearchHit, KifiSearchHit }
-import com.keepit.social.BasicUser
 import org.specs2.mutable.Specification
 import play.api.libs.json._
 
@@ -16,33 +14,6 @@ class SearchAnalyticsTest extends Specification {
   )
 
   "KifiHitContext" should {
-    "be deserialized from a KifiSearchHit" in { // old extensions, to be deprecated
-
-      def check(kifiHitContext: KifiHitContext, kifiSearchHit: KifiSearchHit): Unit = {
-        kifiHitContext.isOwnKeep === kifiSearchHit.isMyBookmark
-        kifiHitContext.isPrivate === kifiSearchHit.isPrivate
-        kifiHitContext.keepCount === kifiSearchHit.count
-        kifiHitContext.keepers === kifiSearchHit.users.map(_.externalId)
-        kifiHitContext.tags === kifiSearchHit.bookmark.collections.toSeq.flatten
-        kifiHitContext.titleMatches === kifiSearchHit.bookmark.titleMatches.length
-        kifiHitContext.urlMatches === kifiSearchHit.bookmark.urlMatches.length
-      }
-
-      val hit1 = BasicSearchHit(title = Some("whatever"), url = "https://whatever", collections = Some(tags), titleMatches = Some(Seq((1, 2), (3, 4))))
-      val kifiSearchHit1 = KifiSearchHit(hit1, 10, false, true, users.map(BasicUser.fromUser), 0.5.toFloat)
-      val kifiHitContext1 = kifiSearchHit1.json.as[KifiHitContext]
-      check(kifiHitContext1, kifiSearchHit1)
-
-      val hit2 = BasicSearchHit(title = None, url = "https://whatever", urlMatches = Some(Seq((1, 2), (3, 4))))
-      val kifiSearchHit2 = KifiSearchHit(hit2, 10, false, true, users.map(BasicUser.fromUser), 0.5.toFloat)
-      val kifiHitContext2 = kifiSearchHit2.json.as[KifiHitContext]
-      check(kifiHitContext2, kifiSearchHit2)
-      val hit3 = BasicSearchHit(title = None, url = "https://whatever")
-      val kifiSearchHit3 = KifiSearchHit(hit3, 10, false, true, users.map(BasicUser.fromUser), 0.5.toFloat)
-      val kifiHitContext3 = kifiSearchHit3.json.as[KifiHitContext]
-      check(kifiHitContext3, kifiSearchHit3)
-      1 === 1
-    }
 
     "be deserialized from current extension payload" in {
       val json = Json.obj(
